requirements.txt
Django == 3.1.2
django-bootstrap4 == 2.2.0
wheel == 0.35.1
coreapi == 2.3.3
Markdown ==  3.2.2
Pygments == 2.6.1
django-filter == 2.3.0
djangorestframework == 3.11.1
drf-yasg == 1.17.1
six~=1.15.0
pillow == 7.2.0
python-dotenv~=0.14.0
PyJWT~=1.7.1
djoser==2.0.3
djangorestframework-simplejwt ==4.4.0
